requirements.txt
Django==4.1.4
django_environ==0.9.0
django_user_agents==0.4.0
environ==1.0
Pillow==9.4.0
python-decouple==3.6
Unidecode==1.3.6
celery==5.2.7
django-celery-beat==2.4.0